Assignment 4/cms_api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
idna==3.10
numpy==2.2.4
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.1
